+++ src/__tests__/gaming/ExtendedNFT721Manager.test.ts
@@ -13,9 +13,10 @@ const fakeTransactionExecuteResponse: ExecuteResponse = {
   transactionResponse: {} as unknown as TransactionResponseExtended
 }
 
+const spyWarn = jest.spyOn(console, 'warn').mockImplementation(() => {})
+
 describe('ExtendedNFT721Manager constructor', () => {
   it('Should instanciate the ExtendedNFT721Manager', () => {
-    const spyWarn = jest.spyOn(console, 'warn')
     const sdk = new ExtendedNFT721Manager()
     expect(sdk).toBeDefined()
     expect(sdk).toBeInstanceOf(ExtendedNFT721Manager)