+++ src/connectors/AbstractWeb3Connector.ts
@@ -3,8 +3,8 @@
 import { BigNumber, BigNumberish, ethers, Signature, Wallet } from 'ethers'
 import errorTypes from '../types/errors'
 import {
-  TransactionResponseExtended,
-  TransactionResponse
+  TransactionResponse,
+  ConnectorResponseExtended
 } from '../types/interfaces'
 import { parseFixed } from '@ethersproject/bignumber'
 import { UnitTypes } from '../types/enums'
@@ -14,7 +14,7 @@ import {
   isRequired,
   matchType
 } from '../utils/validations'
-import { extendTransactionResponse, fromWei } from '../utils/conversions'
+import { connectorResponse, fromWei } from '../utils/conversions'
 import { EIP2612PermitTypedDataSigner } from '../utils/typed-data'
 
 export default abstract class AbstractWeb3Connector {
@@ -133,7 +133,7 @@ export default abstract class AbstractWeb3Connector {
   transferBalance = async (
     amount: string,
     destinationAddress: string
-  ): Promise<TransactionResponseExtended> => {
+  ): Promise<ConnectorResponseExtended> => {
     // doing import inside function to avoid circular dependency
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     if (!this._isActive) {
@@ -146,7 +146,7 @@ export default abstract class AbstractWeb3Connector {
       value: parseFixed(amount, UnitTypes.ETHER)
     }
     const txResponse = await this._account!.sendTransaction(tx)
-    return extendTransactionResponse(txResponse, this)
+    return connectorResponse(txResponse, this)
   }
 
   /**
@@ -305,6 +305,26 @@ export default abstract class AbstractWeb3Connector {
       to: txReceipt.to,
       value: txReceipt.value,
       data: newData,
+      ...speedUp
+    })
+  }
+
+  changeBalanceTransaction = async (
+    txReceipt: TransactionResponse = isRequired('txReceipt'),
+    to?: string | undefined,
+    value?: BigNumber | undefined,
+    gasSpeed?: BigNumber
+  ): Promise<TransactionResponse> => {
+    if (!this._isActive) {
+      throw errorTypes.MUST_ACTIVATE
+    }
+    if (this.isReadOnly) throw errorTypes.READ_ONLY('changeBalanceTransaction')
+
+    const speedUp = await this._speedUpGas(txReceipt, gasSpeed)
+    return this._account!.sendTransaction({
+      from: txReceipt.from,
+      to: to ?? txReceipt.to,
+      value: value ?? txReceipt.value,
       ...speedUp
     })
   }