+++ src/nft/storage/index.ts
@@ -1,0 +1,2 @@
+export { default as ResourceIpfs } from './ResourceIpfs'
+export { default as StorageIpfs } from './StorageIpfs'