+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [1.1.0-dev.6](https://github.com/ripio/sdkjs/compare/v1.1.0-dev.5...v1.1.0-dev.6) (2022-10-31)
+
+
+### Features
+
+* adding ResourceIpfs and StorageIpfs ([0647b9a](https://github.com/ripio/sdkjs/commit/0647b9af9c00dbf0b5e8b9c736e9d410445ed772))
+* changing resource from interface to abstract class ([94faf2c](https://github.com/ripio/sdkjs/commit/94faf2c0a7e972e1f7fb29c27ce1ab31d0290ee9))
+
+# [1.1.0-dev.5](https://github.com/ripio/sdkjs/compare/v1.1.0-dev.4...v1.1.0-dev.5) (2022-10-28)
+
+
+### Features
+
+* add nft class ([d8a73e2](https://github.com/ripio/sdkjs/commit/d8a73e295e813562b5faeb5489004920205dc8ab))
+
+# [1.1.0-dev.4](https://github.com/ripio/sdkjs/compare/v1.1.0-dev.3...v1.1.0-dev.4) (2022-10-27)
+
+
+### Features
+
+* add StorageType and Resourse interfaces ([8606401](https://github.com/ripio/sdkjs/commit/8606401b7e02b618fb9b131453af7e4b9cfe6768))
+* added changeBalanceTransaction on abstract connector to replace a transaction of a transfer balance ([35a4d56](https://github.com/ripio/sdkjs/commit/35a4d56293d1624f35b64478542f2a19a836b970))
+* new interface for transferBalance response, new connectorResponse function for the transferBalance response (with change, speedUp, cancel methods), updated execute response interface (manager is mandatory) ([78814a8](https://github.com/ripio/sdkjs/commit/78814a88a4bb9425034d70ae0b0b30fb41d1cc9c))
+
 # [1.1.0-dev.3](https://github.com/ripio/sdkjs/compare/v1.1.0-dev.2...v1.1.0-dev.3) (2022-10-26)